1. overhaul how geometrical information is read from input file; 2. careful with copying geometries; 3. add support for creating mesh of particles using in-built functions; 4. peridynamics apps works well
